Final_Hackathon/backend/requirements.txt
# Core Framework
fastapi
uvicorn
python-dotenv

# Langchain and AI
langchain
langchain-google-genai

# Database
motor
pymongo

# PDF Generation
fpdf2

# Document Processing
python-docx
PyMuPDF

# Testing
pytest
pytest-asyncio
httpx

# Data Processing
pandas
openpyxl 

langgraph

python-multipart